antsPy/requirements.txt
absl-py==0.11.0
antspynet==0.0
antspyx==0.2.4
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
chart-studio==1.1.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.0
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
nibabel==3.2.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sklearn==0.0
statsmodels==0.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-probability==0.11.1
termcolor==1.1.0
tfp-nightly==0.2.0.dev20201028
threadpoolctl==2.1.0
tifffile==2020.10.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webcolors==1.11.1
Werkzeug==1.0.1
wrapt==1.12.1
